[BUG_FIXED] Fix the plugins shortcuts not working issue. Fix the tooltip on toolbar bug for the plugins icons.

git-svn-id: svn://svn.tuxfamily.org/svnroot/notepadplus/repository/trunk@510 f5eea248-9336-0410-98b8-ebc06183d4e3

+++ PowerEditor/src/Parameters.cpp
@@ -1636,7 +1636,7 @@ void NppParameters::feedShortcut(TiXmlNode *node)
 			int len = (int)_shortcuts.size();
 			for(int i = 0; i < len; i++) 
 			{
-				if (_shortcuts[i].getID() == (unsigned short)id) 
+				if (_shortcuts[i].getID() == (unsigned long)id) 
 				{	//found our match
 					getShortcuts(childNode, _shortcuts[i]);
 					addUserModifiedIndex(i);
